Rearrange code / split out hook
