+++ ui/src/overlap/hooks/useStaffMediaQuery.ts
@@ -1,18 +1,11 @@
 import { useQuery } from "urql";
 import {
-  staffCharacterMediaFragment,
-  staffRoleMediaFragment,
-} from "./queries/staffFragments";
-import {
-  StaffMediaQuery,
   StaffCharacterMediaQueryBatch2,
+  StaffMediaQuery,
   StaffRoleMediaQueryBatch2,
-  staffInfoFragment,
-} from "./queries/staffMedia";
-import { readFragment } from "gql.tada";
-import { useMemo } from "react";
+} from "../queries/staffMedia";
 
-function useStaffMediaQuery(staffId: number) {
+export function useStaffMediaQuery(staffId: number) {
   const [firstBatchResult] = useQuery({
     query: StaffMediaQuery,
     variables: { staffId },
@@ -113,45 +106,4 @@ function useStaffMediaQuery(staffId: number) {
       characterBatch2Result.error ||
       staffBatch2Result.error,
   };
-}
-
-export function Staff(props: { staffId: string }) {
-  const { staffId } = props;
-
-  const queryResult = useStaffMediaQuery(Number.parseInt(staffId));
-
-  const mediaIds = useMemo(() => {
-    return new Set([
-      ...queryResult.data.collapsedCharacters.map(
-        (mediaCharacter) =>
-          readFragment(staffCharacterMediaFragment, mediaCharacter)?.media?.id,
-      ),
-      ...queryResult.data.collapsedStaffRoles.map(
-        (mediaCharacter) =>
-          readFragment(staffRoleMediaFragment, mediaCharacter)?.media?.id,
-      ),
-    ]);
-  }, [queryResult.data]);
-
-  if (queryResult.fetching) {
-    return <>Loading staff information...</>;
-  }
-  if (queryResult.error) {
-    return <>Failed to fetch Staff information.</>;
-  }
-
-  return (
-    <>
-      <p>
-        Page for{" "}
-        {readFragment(staffInfoFragment, queryResult.data?.Staff)?.name?.full}.
-      </p>
-      <p>
-        This person has {queryResult.data?.collapsedCharacters.length} voice
-        roles and {queryResult.data?.collapsedStaffRoles.length} staff roles.
-      </p>
-      <p>{mediaIds.size} total unique Media IDs:</p>
-      <p> {[...mediaIds].join(", ")}</p>
-    </>
-  );
 }
